refactor(core): 🔨 split DKIM/DANE methods into partial classes
